Export chapter titles feature implementation

+++ includes/templates/actions/recreate-cbr/index.php
@@ -12,33 +12,62 @@
 
 
 
-    $cbrBackupDir = $mangaInfo->getOutputDir().'cbr-backup/'.date('Y-m-d-H-i').'/';
 
-    consoleLineInfo('Backing up old .cbr files to '.$cbrBackupDir);
 
-    if(!is_dir($cbrBackupDir)) {
 
-        if(!mkdir($cbrBackupDir,0777,true)) {
-            consoleLineError("Unable to create cbr backup dir [{$cbrBackupDir}] ");
-            exit();
+
+    if($objArgumentsList->shouldKeepCbrBackup()) {
+
+        $cbrBackupDir = $mangaInfo->getOutputDir().'cbr-backup/'.date('Y-m-d-H-i').'/';
+
+        consoleLineInfo('Backing up old .cbr files to '.$cbrBackupDir);
+
+        if(!is_dir($cbrBackupDir)) {
+
+            if(!mkdir($cbrBackupDir,0777,true)) {
+                consoleLineError("Unable to create cbr backup dir [{$cbrBackupDir}] ");
+                exit();
+            }
+
         }
+
+        exec("cp ".$mangaInfo->getCbrDirPath()."/*.cbr {$cbrBackupDir}/");
+
+    } else {
+
+        consoleLineInfo('Removing .existing cbr files ...');
 
     }
 
-    consoleLineInfo('Backing up existing cbr files ...');
-    exec("cp ".$mangaInfo->getCbrDirPath()."/*.cbr {$cbrBackupDir}/");
     exec("rm ".$mangaInfo->getCbrDirPath()."/*.cbr");
+
     consoleLineInfo('Done');
 
     $chapters = $mangaStatus->getAllChaptersList();
 
-    foreach($chapters as $c) {
+    $objChaptersTitles = ChapterTitles::getInstance();
+
+    /**
+     * @var ChapterInfo $c
+     */
+
+    foreach($chapters as &$c) {
+
+        $title = trim($objChaptersTitles->getChapterTitle($c->getNumber()));
+
+        if($title) {
+            $c->setTitle($title);
+        }
 
         $cbrCreator = new CbrCreator(array(
             'mangaInfo' => $mangaInfo,
             'chapterInfo' => $c
         ));
 
-        consoleLineInfo($cbrCreator->createCbr());
+        $cbrCreator->setPrintRarOutput(false);
+
+        $cbrCreator->createCbr();
 
     }
+
+    $mangaStatus->updateAllChaptersList($chapters);
